+++ core/src/net/pixelstatic/codetesting/modules/ecstesting/components/RenderComponent.java
@@ -1,18 +1,18 @@
 package net.pixelstatic.codetesting.modules.ecstesting.components;
 
-import net.pixelstatic.codetesting.modules.ecstesting.Entity;
-import net.pixelstatic.codetesting.modules.ecstesting.EntityTester;
-
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.math.MathUtils;
+
+import net.pixelstatic.codetesting.modules.ecstesting.Entity;
+import net.pixelstatic.codetesting.modules.ecstesting.EntityTester;
 
 public class RenderComponent implements Component{
 	float x, y;
 	
 	public void update(Entity entity, EntityTester tester){
-		tester.batch.draw(tester.atlas.findRegion("repulsemagnet"), (float)entity.getValue("x"), (float)entity.getValue("y"));
-		entity.incrementValue("x", MathUtils.sin((float)entity.getValue("y")/10f));
-		entity.incrementValue("y", MathUtils.sin((float)entity.getValue("x")/10f));
+		tester.batch.draw(tester.atlas.findRegion("repulsemagnet"), (float)(Float)entity.getValue("x"), (float)(Float)entity.getValue("y"));
+		entity.incrementValue("x", MathUtils.sin((float)(Float)entity.getValue("y")/10f));
+		entity.incrementValue("y", MathUtils.sin((float)(Float)entity.getValue("x")/10f));
 	}
 	
 	public void init(Entity entity){
